requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.1
chromadb==1.5.5
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
gunicorn==22.0.0
